Extract Vanity::Adapters to separate file.
